+++ test/test_parser.py
@@ -211,9 +211,9 @@ def test_parser_from_isbn():
     entry = list(entries.values())[0]
     assert entry.label == 'Lutz2013'
     assert entry.data['author'] == 'Mark Lutz'
-    assert entry.data['pages'] == 1540
+    assert entry.data['pages'] == '1540'
     assert entry.data['title'] == 'Learning Python'
-    assert entry.data['year'] == 2013
+    assert entry.data['year'] == '2013'
 
 
 def test_parser_from_arxiv():