+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- ISBN parser was missing the ENTRYTYPE and did not use strings for number fields
+
 
 ## [2.4.0] - 2020-10-14